requirements.txt
WTForms~=3.0.0
Flask~=2.0.2
setuptools~=57.0.0
python-dotenv~=0.19.2
Werkzeug~=2.0.2
SQLAlchemy~=1.4.27